tests

+++ src/MVC/controllers/FileTypesController.java
@@ -1,22 +1,33 @@
 package MVC.controllers;
 
 import MVC.models.FileType;
+import MVC.models.User;
 
 public class FileTypesController {
-    public boolean create(String _title){
-        FileType file_type = new FileType(_title);
-        if(file_type != null){
-            return true;
+    public static boolean create(int _user_id, String _title){
+        User user = User.find(_user_id);
+        if(user.has_role("admin")){
+            FileType file_type = new FileType(_title);
+            if(file_type != null){
+                return true;
+            } else {
+                return false;
+            }
         } else {
             return false;
         }
     }
 
-    public boolean delete(int _id){
+    public static boolean delete(int _id, int _user_id){
+        User user = User.find(_user_id);
         FileType type = FileType.find(_id);
         if(type != null){
-            type.delete();
-            return true;
+            if(user.has_role("admin")) {
+                type.delete();
+                return true;
+            } else {
+                return false;
+            }
         } else {
             return false;
         }
